@@ -1,3 +1,3 @@
 module ActiveRest
-  VERSION = '6.7.0'
+  VERSION = '6.8.0'
 end
